test: extract mockPaidTemplate into TemplateHelper
Addresses review feedback: the paid-template mock is reusable across
specs, so it belongs in the shared helper rather than duplicated inline.
Also trims comments that restate the assertion.
