+++ browser_tests/tests/templates/partnerNodesEducationCard.spec.ts
@@ -1,45 +1,22 @@
 import { expect } from '@playwright/test'
-import type { Page } from '@playwright/test'
 
-import { makeTemplate } from '@e2e/fixtures/data/templateFixtures'
 import { comfyPageFixture as test } from '@e2e/fixtures/ComfyPage'
-import type { TemplateHelper } from '@e2e/fixtures/helpers/TemplateHelper'
-import {
-  createTemplateHelper,
-  withTemplates
-} from '@e2e/fixtures/helpers/TemplateHelper'
+import { mockPaidTemplate } from '@e2e/fixtures/helpers/TemplateHelper'
 import { TestIds } from '@e2e/fixtures/selectors'
 
 const PAID_TEMPLATE = 'paid-template'
-
-async function mockPaidTemplate(page: Page): Promise<TemplateHelper> {
-  const templates = createTemplateHelper(
-    page,
-    withTemplates([
-      makeTemplate({
-        name: PAID_TEMPLATE,
-        title: 'Paid Template',
-        description: 'Uses partner nodes.',
-        openSource: false
-      })
-    ])
-  )
-  await templates.mock()
-  // The paid template's workflow genuinely contains a partner node; the
-  // card additionally gates on live graph content, not just the flag.
-  await templates.mockWorkflow(
-    PAID_TEMPLATE,
-    'browser_tests/assets/partner_api_node.json'
-  )
-  return templates
-}
+const PARTNER_WORKFLOW = 'browser_tests/assets/partner_api_node.json'
 
 test.describe('Partner nodes education card (local)', () => {
   test('shows on paid template load, hides on graph switch, re-shows without a seen-flag', async ({
     comfyPage
   }) => {
     const page = comfyPage.page
-    const templates = await mockPaidTemplate(page)
+    const templates = await mockPaidTemplate(
+      page,
+      PAID_TEMPLATE,
+      PARTNER_WORKFLOW
+    )
 
     const card = page.getByTestId(TestIds.partnerNodes.educationCard)
 
@@ -47,17 +24,17 @@ test.describe('Partner nodes education card (local)', () => {
     await expect(card).toBeVisible()
     await expect(card).toContainText('See the difference? Drag to compare.')
 
-    // Switching to a graph without partner nodes hides the card without
-    // requiring a dismissal — it must not outlive the template it describes.
+    // Leaving the graph retires the card — it must not describe a workflow the
+    // user has left.
     await comfyPage.workflow.loadWorkflow('default')
     await expect(card).toHaveCount(0)
 
-    // Dismissing carries no seen-flag: the next paid load shows it again.
     await templates.load(PAID_TEMPLATE)
     await expect(card).toBeVisible()
     await page.getByTestId(TestIds.partnerNodes.educationCardDismiss).click()
     await expect(card).toHaveCount(0)
 
+    // No seen-flag: the next paid load shows it again.
     await templates.load(PAID_TEMPLATE)
     await expect(card).toBeVisible()
   })
